new web pageee/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.0
sqlite3
hashlib
datetime
json
re
threading
time
logging 